requirements.txt
ansi2html==1.8.0
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
category-encoders==2.6.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.5
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Cython==3.0.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
Django==4.2.3
django-cors-headers==4.1.0
django-password-validators==1.7.1
django-rest==0.8.7
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
drf-yasg==1.21.7
enum34==1.1.10
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
fonttools==4.41.0
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
inflection==0.5.1
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kaleido==0.2.1
kiwisolver==1.4.4
lightgbm==4.0.0
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
nbformat==5.9.1
nest-asyncio==1.5.6
numba==0.57.1
numpy==1.23.5
openapi-codec==1.3.2
orjson==3.9.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
plotly==5.15.0
plotly-resampler==0.9.1
pmdarima==2.0.3
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycaret==3.0.4
Pygments==2.15.1
PyJWT==2.7.0
pyod==1.1.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
retrying==1.3.4
schemdraw==0.15
scikit-base==0.5.0
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.11.1
simplejson==3.19.1
six==1.16.0
sktime==0.20.1
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
tbats==1.1.3
tenacity==8.2.2
threadpoolctl==3.2.0
tornado==6.3.2
tqdm==4.65.0
trace-updater==0.0.9.1
traitlets==5.9.0
tsdownsample==0.1.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wrapt==1.15.0
wurlitzer==3.0.3
xxhash==3.2.0
yellowbrick==1.5
zipp==3.16.2
